Move classes to respective assemblies
