+++ BluetoothChargeCheck.Models/BluetoothDeviceData.cs
