+++ BluetoothChargeCheck.Resources/SegoeFluentIcons.ttf
