--- BluetoothChargeCheck.ViewModels/Class1.cs
@@ -1,10 +1,0 @@
-// <copyright file = "Class1.cs" company = "Digital Cloud Technologies">
-// Copyright (c) Digital Cloud Technologies.All rights reserved.
-// </copyright>
-
-namespace DCT.BluetoothChargeCheck.ViewModels;
-
-public class Class1
-{
-}
-
